+++ data/src/main/kotlin/com/acon/data/repository/SpotRepositoryImpl.kt
@@ -52,7 +52,7 @@ class SpotRepositoryImpl @Inject constructor(
     }
 
     override suspend fun getSpotMenuList(
-        spotId: Int
+        spotId: Long
     ): Result<List<SpotDetailMenu>> {
         return runCatchingWith(*GetSpotMenuListError.createErrorInstances()) {
             spotRemoteDataSource.getSpotMenuList(spotId).menuList.map { it.toSpotDetailMenu() }
